+++ web/routers/router.go
@@ -23,7 +23,6 @@ func Init() {
 			beego.NSAutoRouter(&controllers.AuthController{}),
 			beego.NSAutoRouter(&controllers.GlobalController{}),
 			beego.NSAutoRouter(&controllers.ProxyNodeController{}),
-			beego.NSAutoRouter(&controllers.OmniProxyController{}),
 		)
 		beego.AddNamespace(ns)
 	} else {
@@ -34,6 +33,5 @@ func Init() {
 		beego.AutoRouter(&controllers.AuthController{})
 		beego.AutoRouter(&controllers.GlobalController{})
 		beego.AutoRouter(&controllers.ProxyNodeController{})
-		beego.AutoRouter(&controllers.OmniProxyController{})
 	}
 }